Munge SGB colors. Add GBC colors. #245
Firstly: it was discovered (or rather, noticed) that the sprites in
Green have a greenish background (technically, first palette entry)
rather than a reddish one.

1. To properly represent this difference between Red and Green, we would
   have to give Green its own directory.  This is silly, considering
   that the sprites are identical in every other respect.
2. The background is not something inherent to the sprites, but a tint
   which pervades the different versions of the game: Red is red, Green
   is green, and Yellow is yellow.  (Blue is red, not blue.  Don't ask
   me why.)

For these reasons, the Gen I sprites have had their background changed
to white.  (The red background was ugly anyway.)

Secondly: Yellow has an additional set of palettes, which it uses when
played on a Gameboy Color.  Sprites colored with these palettes have
been added to media/yellow/gbc and media/yellow/back/gbc.

It was suggested that the GBC-paletted sprites should be considered the
"main" ones; the argument being that the GBC is more canon (because it
is the platform most people would have played on).

I have kept the SGB-paletted sprites as the main ones, for the following
reasons:

1. By the above argument, the *grayscale* sprites should be considered
   the main ones for Red & Green & Blue, since that is how they would
   appear on a Gameboy.
2. The SGB colors go better with the colors we have for Red & Green &
   Blue.  (They should--they were designed to.)