(Re)ripped GSC sprites, sans animations. #197 #102
We now have backsprites and still-frame sprites for Crystal.  I'm
considering the still-frame sprites to be the canon ones to match
Emerald; the current (backwards) animated ones have been moved to
animated/.  We also now have separate backsprites for Silver even
though they're the same as Gold.  Crystal's do sometimes differ.

Turns out the sprites are actually all different sizes, too; it's
just that nobody can tell in-game, since the game places them all
in the same frame.  (Actually, I think it might bottom-align them
in battle and center them in the Pokédex or something; not sure.)
The backsprites are all the same size, but it's not the same size
as the ones we had before.  Either way, I think it's more correct
not to pad the sprites to a uniform size, in-game display aside.

I converted my script's PPM output to PNG with OptiPNG.  It seems
to have rounded the five-bit-per-channel colors to fit eight bits
per channel, which I like.  (Rather than flooring or something?)