Replace mp3 cries with newly-ripped ogg cries. #91
Created with oggenc, using the default quality of 3.

As with sprites, Shaymin has its default form's cry both as 492.ogg, so
that one can ignore anything that isn't ###.ogg and get a cry for every
Pokémon, and as 492-land.ogg, to avoid needing "don't append the form if
it's the default" junk.