B/W: Footprints!  Replace the old ones with a B/W rip.
The old ones were 3× for some reason and the blank ones all had Gen II's
"NONE" image.  Also, the first 386 had their Gen III footprints; a
handful matched the actual feet badly enough that they changed in D/P.
See #250.

And, of course, this means we aren't missing footprints for the new
Pokémon.

There was also an extra footprint for Cherrim before, even though it was
the same...?